Fix onRemoveLoc

+++ js/app.controller.js
@@ -84,17 +84,23 @@ function renderLocs(locs) {
 }
 
 function onRemoveLoc(locId) {
-    locService.remove(locId)
-        .then(confirm('Are you sure you want to remove this location?'))
-        .then(() => {
-            flashMsg('Location removed')
-            unDisplayLoc()
-            loadAndRenderLocs()
-        })
-        .catch(err => {
-            console.error('OOPs:', err)
-            flashMsg('Cannot remove location')
-        })
+    const isAgree = confirm('Are you sure you want to remove this location?')
+    if (isAgree) {
+        locService.remove(locId)
+            .then(() => {
+                flashMsg('Location removed')
+                unDisplayLoc()
+                loadAndRenderLocs()
+            })
+            .catch(err => {
+                console.error('OOPs:', err)
+                flashMsg('Cannot remove location')
+            })
+
+    } else {
+        flashMsg('This item will not be deleted - as you wanted.')
+    }
+
 }
 
 function onSearchAddress(ev) {
@@ -199,13 +205,13 @@ function displayLoc(loc) {
 
     mapService.panTo(loc.geo)
     mapService.setMarker(loc)
-    const latLng = {lat:loc.geo.lat, lng: loc.geo.lng}
-    
+    const latLng = { lat: loc.geo.lat, lng: loc.geo.lng }
+
 
     const el = document.querySelector('.selected-loc')
     el.querySelector('.loc-name').innerText = loc.name
     el.querySelector('.loc-address').innerText = loc.geo.address
-    el.querySelector('.loc-distance-fromMyPos').innerText = 'Distance: '+utilService.getDistance(gUserPos, latLng, 'K') + ' KM'
+    el.querySelector('.loc-distance-fromMyPos').innerText = 'Distance: ' + utilService.getDistance(gUserPos, latLng, 'K') + ' KM'
     el.querySelector('.loc-rate').innerHTML = '★'.repeat(loc.rate)
     el.querySelector('[name=loc-copier]').value = window.location
     el.classList.add('show')
